+++ lab3/z3/ll_lazy.c
@@ -181,7 +181,7 @@ int ll_is_sorted(ll_t *ll) {
 
 unsigned long long ll_length(ll_t *ll) {
     ll_node_t *curr = ll->head;
-    int ret = 0;
+    unsigned long long ret = 0;
 
     curr = curr->next;
 
@@ -195,7 +195,7 @@ unsigned long long ll_length(ll_t *ll) {
 
 unsigned long long ll_sum_of_keys(ll_t *ll) {
     ll_node_t *curr = ll->head;
-    int ret = 0;
+    unsigned long long ret = 0;
 
     curr = curr->next;
